feat: added side queue position setting
The next piece queue on the side when positioned near the top is hard to
see when the eyes need to stay focused on the grid at high gravity.

Perhaps the setting could be expanded to center the queue vertically on
the side depending on how many pieces are visible in the queue in a
given game mode.